Swap order of optional group for mariadb banners
MariaDB typically uses the 10.x.x version for anything post 5.5.5 (or
so?), but tends to include the 5.5.5 version in the beginning of the
banner for backwards compatibility (maybe?). Since the 10.x.x version
varies while the 5.5.5 version remains static, capture that instead and
treat the 5.5.5 as optional/non-captured.

Actual 5.x.x versions still capture correctly when no 10.x.x
version is present in the banner.