Improve Exim coverage, add examples
The primary purpose of this commit is to improve coverage of Exim in smtp_banners.xml.

In addition to this I've added multiple examples for some fingerprints that were missing them. During this I found mulitple fingerprints that didn't match the examples found in the description and/or in the wild.  These were fixed as well.