Dramatically increase performance of updating CPEs
The previous method had to parse the CPE validity file for each input
fingerprint file to generate CPEs for.

Instead, pass the script a list of filenames via stdin and parse the
file only once.

Remove the suggestion of using `parallel` since it won't increase
performance anymore (and doing so is not necessary)