Break Windows 10 Mobile fingerprint in two
Not sure the best way to handle thing. Fingerprints with optional
fields that are needed to sub into a generated CPE will build malformed
CPEs.

This causes this FPs validation tests to fail because that happens.

It seems like a decent solution would be adding some type of "optional"
attribute to the extracted fingerprint parameters. The CPE generator
would then treat those fields as wildcards at generation time, or skip
generating CPEs all-together. I don't want to go ahead and implement
that, though; it feels like this patch is getting big.