Fix ReDos regex expression
The Regex changed suffers from an exponential degree of ambiguity.

As a result it is trivial to provide an input string which takes
an incredibly long time to match (due to exponential growth by length
of input string).

Depending on the matching engine used, this can lead to a DoS of a service.

More details: https://owasp.org/www-community/attacks/Regular_expression_Denial_of_Service_-_ReDoS